MRKT-200: Refactor studio to use the shared claim earnings components
